examples/requirements/test.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml examples/requirements/test.in --all-extras -o examples/requirements/test.txt
alabaster==0.7.16
    # via sphinx
anyio==4.9.0
    # via
    #   starlette
    #   watchfiles
asgiref==3.8.1
    # via django
asttokens==3.0.0
    # via stack-data
attrs==25.3.0
    # via
    #   outcome
    #   trio
babel==2.17.0
    # via sphinx
backports-tarfile==1.2.0
    # via jaraco-context
beautifulsoup4==4.13.4
    # via
    #   -r examples/requirements/test.in
    #   ska (pyproject.toml)
build==1.2.2.post1
    # via ska (pyproject.toml)
cachetools==6.0.0
    # via tox
certifi==2025.4.26
    # via
    #   requests
    #   selenium
cffi==1.17.1
    # via cryptography
chardet==5.2.0
    # via tox
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   pydoclint
    #   uvicorn
colorama==0.4.6
    # via
    #   sphinx-autobuild
    #   tox
coverage==7.8.2
    # via
    #   -r examples/requirements/test.in
    #   ska (pyproject.toml)
    #   pytest-cov
cryptography==45.0.3
    # via
    #   pyjwt
    #   secretstorage
decorator==5.2.1
    # via ipython
detect-secrets==1.5.0
    # via ska (pyproject.toml)
distlib==0.3.9
    # via virtualenv
django==5.2.1
    # via
    #   django-nine
    #   djangorestframework
    #   drf-jwt
django-constance==4.3.2
    # via ska (pyproject.toml)
django-nine==0.2.7
    # via ska (pyproject.toml)
djangorestframework==3.16.0
    # via
    #   ska (pyproject.toml)
    #   drf-jwt
doc8==1.1.2
    # via ska (pyproject.toml)
docstring-parser-fork==0.0.12
    # via pydoclint
docutils==0.19
    # via
    #   doc8
    #   readme-renderer
    #   restructuredtext-lint
    #   sphinx
    #   sphinx-no-pragma
    #   sphinx-rtd-theme
drf-jwt==1.19.2
    # via ska (pyproject.toml)
exceptiongroup==1.3.0
    # via
    #   anyio
    #   ipython
    #   pytest
    #   trio
    #   trio-websocket
executing==2.2.0
    # via stack-data
factory-boy==3.3.3
    # via -r examples/requirements/test.in
fake-py==0.11.8
    # via ska (pyproject.toml)
faker==37.3.0
    # via
    #   -r examples/requirements/test.in
    #   factory-boy
filelock==3.18.0
    # via
    #   tox
    #   virtualenv
h11==0.16.0
    # via
    #   uvicorn
    #   wsproto
id==1.5.0
    # via twine
idna==3.10
    # via
    #   anyio
    #   requests
    #   trio
imagesize==1.4.1
    # via sphinx
importlib-metadata==8.7.0
    # via keyring
iniconfig==2.1.0
    # via pytest
ipython==8.37.0
    # via ska (pyproject.toml)
jaraco-classes==3.4.0
    # via keyring
jaraco-context==6.0.1
    # via keyring
jaraco-functools==4.1.0
    # via keyring
jedi==0.19.2
    # via ipython
jeepney==0.9.0
    # via
    #   keyring
    #   secretstorage
jinja2==3.1.6
    # via sphinx
keyring==25.6.0
    # via twine
libcst==1.8.0
    # via monkeytype
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
matplotlib-inline==0.1.7
    # via ipython
mdurl==0.1.2
    # via markdown-it-py
mock==5.2.0
    # via
    #   -r examples/requirements/test.in
    #   ska (pyproject.toml)
monkeytype==23.3.0
    # via pytest-monkeytype
more-itertools==10.7.0
    # via
    #   jaraco-classes
    #   jaraco-functools
mypy==1.16.0
    # via ska (pyproject.toml)
mypy-extensions==1.1.0
    # via
    #   monkeytype
    #   mypy
nh3==0.2.21
    # via readme-renderer
outcome==1.3.0.post0
    # via
    #   trio
    #   trio-websocket
packaging==25.0
    # via
    #   build
    #   django-nine
    #   pyproject-api
    #   pytest
    #   sphinx
    #   tox
    #   twine
parso==0.8.4
    # via jedi
pathspec==0.12.1
    # via mypy
pbr==6.1.1
    # via stevedore
pexpect==4.9.0
    # via ipython
platformdirs==4.3.8
    # via
    #   tox
    #   virtualenv
pluggy==1.6.0
    # via
    #   pytest
    #   tox
prompt-toolkit==3.0.51
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
py==1.11.0
    # via -r examples/requirements/test.in
pycparser==2.22
    # via cffi
pydoclint==0.6.6
    # via ska (pyproject.toml)
pygments==2.19.1
    # via
    #   doc8
    #   ipython
    #   pytest
    #   readme-renderer
    #   rich
    #   sphinx
pyjwt==2.10.1
    # via drf-jwt
pyproject-api==1.9.1
    # via tox
pyproject-hooks==1.2.0
    # via build
pysocks==1.7.1
    # via urllib3
pytest==8.4.0
    # via
    #   -r examples/requirements/test.in
    #   ska (pyproject.toml)
    #   pytest-codeblock
    #   pytest-cov
    #   pytest-django
    #   pytest-monkeytype
    #   pytest-ordering
pytest-codeblock==0.1.8
    # via
    #   -r examples/requirements/test.in
    #   ska (pyproject.toml)
pytest-cov==6.1.1
    # via -r examples/requirements/test.in
pytest-django==4.11.1
    # via
    #   -r examples/requirements/test.in
    #   ska (pyproject.toml)
pytest-monkeytype==1.1.0
    # via -r examples/requirements/test.in
pytest-ordering==0.6
    # via -r examples/requirements/test.in
pyyaml==6.0.2
    # via
    #   detect-secrets
    #   libcst
readme-renderer==43.0
    # via twine
requests==2.32.3
    # via
    #   detect-secrets
    #   id
    #   requests-toolbelt
    #   sphinx
    #   twine
requests-toolbelt==1.0.0
    # via twine
restructuredtext-lint==1.4.0
    # via doc8
rfc3986==2.0.0
    # via twine
rich==14.0.0
    # via twine
ruff==0.11.12
    # via ska (pyproject.toml)
secretstorage==3.3.3
    # via keyring
selenium==4.33.0
    # via -r examples/requirements/test.in
setuptools==80.9.0
    # via pbr
sniffio==1.3.1
    # via
    #   anyio
    #   trio
snowballstemmer==3.0.1
    # via sphinx
sortedcontainers==2.4.0
    # via trio
soupsieve==2.7
    # via
    #   -r examples/requirements/test.in
    #   ska (pyproject.toml)
    #   beautifulsoup4
sphinx==5.3.0
    # via
    #   ska (pyproject.toml)
    #   sphinx-autobuild
    #   sphinx-no-pragma
    #   sphinx-rtd-theme
    #   sphinxcontrib-jquery
sphinx-autobuild==2024.10.3
    # via ska (pyproject.toml)
sphinx-no-pragma==0.1.2
    # via ska (pyproject.toml)
sphinx-rtd-theme==2.0.0
    # via ska (pyproject.toml)
sphinxcontrib-applehelp==2.0.0
    # via sphinx
sphinxcontrib-devhelp==2.0.0
    # via sphinx
sphinxcontrib-htmlhelp==2.1.0
    # via sphinx
sphinxcontrib-jquery==4.1
    # via sphinx-rtd-theme
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==2.0.0
    # via sphinx
sphinxcontrib-serializinghtml==2.0.0
    # via sphinx
sqlparse==0.5.3
    # via django
stack-data==0.6.3
    # via ipython
starlette==0.47.0
    # via sphinx-autobuild
stevedore==5.4.1
    # via doc8
tomli==2.2.1
    # via
    #   build
    #   coverage
    #   doc8
    #   mypy
    #   pydoclint
    #   pyproject-api
    #   pytest
    #   tox
tox==4.26.0
    # via -r examples/requirements/test.in
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
trio==0.30.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.12.2
    # via selenium
twine==6.1.0
    # via ska (pyproject.toml)
typing-extensions==4.13.2
    # via
    #   anyio
    #   asgiref
    #   beautifulsoup4
    #   exceptiongroup
    #   ipython
    #   mypy
    #   rich
    #   selenium
    #   tox
    #   uvicorn
tzdata==2025.2
    # via faker
urllib3==2.4.0
    # via
    #   requests
    #   selenium
    #   twine
uv==0.7.9
    # via ska (pyproject.toml)
uvicorn==0.34.3
    # via sphinx-autobuild
virtualenv==20.31.2
    # via tox
watchfiles==1.0.5
    # via sphinx-autobuild
wcwidth==0.2.13
    # via prompt-toolkit
websocket-client==1.8.0
    # via selenium
websockets==15.0.1
    # via sphinx-autobuild
wheel==0.45.1
    # via ska (pyproject.toml)
wsproto==1.2.0
    # via trio-websocket
zipp==3.22.0
    # via importlib-metadata
